feat: add solve mode for analyzing game positions

+++ datagen/src/main.rs
@@ -36,7 +36,7 @@ enum SubCommands {
         records_per_file: u32,
 
         #[arg(long, default_value = "128")]
-        hash_size: i32,
+        hash_size: usize,
 
         #[arg(long, default_value = "12")]
         level: usize,
